Add instance variable to Tree class
This commit adds the @root variable class to the Tree class. The root
variable is assigned to be the returned value of the build_tree method
which is the root of the given array.

The insert class method is also added but unfinished. The idea is to
compare the value to the root of the given array and then determine if
it needs to be compared to the left or right subtree.