Add @root Tree class initialize
This commit adds the @root instance variable to the Tree class
initialize method. This variable is equal to the returned node of
the build_tree class method.

The insert method is still incomplete, however, the pretty_print method
was added to help visualize whether the build_tree actually works. In
order for this method to work I had to change '@left_sub' and
'@right_sub' instance variables in the Node class(and anywhere else it
references) to be 'left' and 'right'.