Tree class method insert
This commit adds the insert method inside Tree class, however there is
an error. I am trying to add the value, 50, which is currently not in
the tree but I get the message that is triggered when
value == root.value.

The method parameters are 'value' and 'root' with 'root' set to @root by
default.