Tree class insert method
This commit continues adding to the insert class method. Currently I
keep running in to a SystemStackError because the call stack is too deep.
This indicates that the recursive case keeps going and I need to the
adjust the base case.

Right now the two small if blocks before the recursive block is trying
to see if the value equals the root value or if the root value is nil.
The second if block can be removed since it is already written into the
ternary operator statements below. This means I need to reconfigure my
base case in order to stop the recursion from happening in an endless
loop.