Tree class insert method - case statements
This commit adds a case statment block to solve the insert method. The
flow of the block seems logical to me and looks much more concise when
compared to the 'if..elsif' block below it.

However, using the case statements only produces a tree without the
value added. Using the 'if..elsif' block adds the new node value but
only that node value exists in the tree, all others don't show up when
pretty_print is called.