Tree class insert method complete
This commit completes the insert method to add a node to the tree.

Initially I kept getting either a SystemStackError, a tree without the
added node or a tree with just the node itself when using the case
statement block. I went back to using an if..else block after seeing
others' work. I think the main reason why the method works now is
because right after determining if the root.value is less than or
greater than the value to add, it checks to see if root.right or
root.left is nil and adds a new Node before calling the recursive
statement.