Tree class delete method
This commit adds the Tree class method, delete.

The current working pseudocode I want to work off of is:

  * First, traverse the tree until the value to be deleted is found.
  * Once value == root.value, assess how many child nodes it has.
    - If there are no child nodes - delete it.
    - If there is one child node - copy the value of the child node to
      the node that is to be deleted.
    - If there are two child nodes - (will work out this case later).

I still need to figure out what the base case is and where the recursive
statement will be placed.

Currently, I am able to delete the value of a single node tree. In the
case of a two-node tree, I can delete the parent node but the child node
needs to be repositioned to take the parent nodes place.