Reconfigure delete method as if..elsif
This commit changes the format of the delete class method to be an
if..elsif block of code. The recursive statements are called if the
value is greater than or less than the root.value and the root is
reassigned to be root.left or root.right depending on the comparison
between the two arguments.

The method doesn't seem to work correctly but it looks like it is a bit
closer to the solution than before.