Tree class method level_order
This commit adds the level_order method. Currently I am getting a
NoMethodError when the code runs into root.left/root.right nodes. The
pseudocode may need some adjusting but the general steps seem to be the
right idea.

As of now, the method only prints out the very first root of the given
array.