Tree class method, level_order
This commit changes the code block for the revised pseudocode in the
previous commit. The code starts as a loop that breaks if the queue
array doesn't have any elements inside. Currently the code seems to
display the correct order of node values when working with a three
element array but it prints it out three times. I suspect it may be due
to the two recursive statements inside the loop.