Tree class method, delete, node with two children
This commit adds the block to handle deletion of a node with two child
nodes. It deletes the correct node but there is still an issue with
sorting out the rest of the tree after deletion.

When deleting the value, 67, from the array and calling pretty_print,
there ends up being two nodes with the value 9. One of the 9's is in
the correct space but the second one should be occupied by the 23 that
it is tethered to.